@@ -11,10 +11,10 @@
   },
   "dependencies": {
     "about-window": "1.8.0",
-    "react": "16.2.0",
-    "react-dom": "16.2.1",
+    "prop-types": "^15.7.2",
+    "react": "16.3.0",
+    "react-dom": "16.3.0",
     "react-redux": "5.0.6",
-    "react-router-dom": "^4.2.2",
     "redux": "^3.7.2",
     "uuid": "3.1.0"
   },
