@@ -6,6 +6,7 @@ import { addFreehandStream, addCircularStream, addLinearStream, updateStreamPosi
 import { updateFPSCount } from '../../actions/Transport';
 import { bindActionCreators } from 'redux';
 import { calculateDistance, getPosition, calculateNodeBorderDistance, timestamp, getNodeById, getNodesWithinDistance } from '../../utils/utils';
+import PropTypes from 'prop-types';
 import {
   addMidiNode,
   selectNode,
@@ -56,6 +57,8 @@ class Canvas extends React.Component {
     this.backgroundX = 0;
     this.backgroundY = 0;
 
+    this.canvasRef = React.createRef();
+
     this.state = {
       mouseDown: false,
       width: props.app.width - config.controlPanel.width - config.menu.width,
@@ -64,16 +67,16 @@ class Canvas extends React.Component {
   }
   
   componentDidMount() {
-    this.canvasContext = this.refs.canvas.getContext('2d');
+    this.canvasContext = this.canvasRef.current.getContext('2d');
     this.draw();
     this.flow();
   }
 
-  componentWillReceiveProps(nextProps) {
-    this.setState({
+  static getDerivedStateFromProps(nextProps) {
+    return {
       width: nextProps.app.width - config.controlPanel.width - config.menu.width,
       height: nextProps.app.height - config.transport.height
-    });
+    };
   }
 
   onMouseDown(event) {
@@ -450,13 +453,15 @@ class Canvas extends React.Component {
   }
 
   render() {
+    const { width, height, mouseDown } = this.state;
+
     let canvasStyle = {
       cursor: 'crosshair',
       backgroundPosition: this.backgroundX + 'px ' + this.backgroundY + 'px'
     };
 
     if (this.props.devices.grab) {
-      if (this.state.mouseDown) {
+      if (mouseDown) {
         canvasStyle.cursor = '-webkit-grabbing';
       } else {
         canvasStyle.cursor = '-webkit-grab';
@@ -467,9 +472,9 @@ class Canvas extends React.Component {
       <canvas
         className="canvas"
         draggable="true"
-        ref="canvas"
-        width={this.state.width}
-        height={this.state.height}
+        ref={this.canvasRef}
+        width={width}
+        height={height}
         onMouseMove={this.onMouseMove}
         onMouseDown={this.onMouseDown}
         onMouseUp={this.onMouseUp}
@@ -481,13 +486,14 @@ class Canvas extends React.Component {
 }
 
 const mapStateToProps = (state) => {
+  const { devices, nodes, streams, collisions, transport, app } = state;
   return {
-    devices: state.devices,
-    nodes: state.nodes,
-    streams: state.streams,
-    collisions: state.Collisions,
-    transport: state.transport,
-    app: state.app
+    devices,
+    nodes,
+    streams,
+    collisions,
+    transport,
+    app,
   };
 };
 
@@ -512,6 +518,33 @@ const mapDispatchToProps = (dispatch) => {
     updateFPSCount: bindActionCreators(updateFPSCount, dispatch),
     deselectStreams: bindActionCreators(deselectStreams, dispatch)
   };
+};
+
+Canvas.propTypes = {
+  app: PropTypes.object,
+  devices: PropTypes.object,
+  streams: PropTypes.array,
+  nodes: PropTypes.array,
+  transport: PropTypes.object,
+  updateFPSCount: PropTypes.func,
+  onMouseMove: PropTypes.func,
+  deselectStreams: PropTypes.func,
+  addFreehandStream: PropTypes.func,
+  deselectNodes: PropTypes.func,
+  addCircularStream: PropTypes.func,
+  addLinearStream: PropTypes.func,
+  updateSelectedNodePositionByDelta: PropTypes.func,
+  updateNodePositionByDelta: PropTypes.func,
+  updateStreamPositionByDelta: PropTypes.func,
+  addMidiNode: PropTypes.func,
+  selectNode: PropTypes.func,
+  cloneNode: PropTypes.func,
+  unlinkNodes: PropTypes.func,
+  linkNodes: PropTypes.func,
+  playNode: PropTypes.func,
+  stopNode: PropTypes.func,
+  enqueueParticle: PropTypes.func,
+  dequeueParticle: PropTypes.func,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Canvas);
