@@ -1,12 +1,12 @@
 export const hydrationStarted = () => {
   return {
-    type: 'HYDRATION_STARTED'
+    type: 'HYDRATION_STARTED',
   };
 };
 
 export const hydrationComplete = () => {
   return {
-    type: 'HYDRATION_COMPLETE'
+    type: 'HYDRATION_COMPLETE',
   };
 };
 
@@ -14,13 +14,13 @@ export const setWindowSize = (width, height) => {
   return {
     type: 'SET_WINDOW_SIZE',
     width,
-    height
+    height,
   };
 };
 
 export const setDirtyStatus = (status) => {
   return {
     type: 'SET_DIRTY_STATUS',
-    status
+    status,
   };
 };
