@@ -1,6 +1,6 @@
 export const toggleDevice = (device) => {
   return {
     type: 'TOGGLE_DEVICE',
-    device
+    device,
   };
 };
