@@ -1,6 +1,6 @@
 export const addDestination = (destination) => {
   return {
     type: 'ADD_DESTINATION',
-    destination
+    destination,
   };
 };
