@@ -3,31 +3,7 @@ export const addMidiNode = (position, relay=false, id=null) => {
     type: 'ADD_MIDI_NODE',
     position,
     relay,
-    id
-  };
-};
-
-export const setNodeOsc1Frequency = (id, frequency) => {
-  return {
-    type: 'SET_NODE_OSC1_FREQUENCY',
     id,
-    frequency
-  };
-};
-
-export const setNodeOsc2Frequency = (id, frequency) => {
-  return {
-    type: 'SET_NODE_OSC2_FREQUENCY',
-    id,
-    frequency
-  };
-};
-
-export const setNodeVolume = (id, volume) => {
-  return {
-    type: 'SET_NODE_VOLUME',
-    id,
-    volume
   };
 };
 
@@ -36,90 +12,59 @@ export const setNodeSource = (id, buffer, path) => {
     type: 'SET_NODE_SOURCE',
     id,
     buffer,
-    path
+    path,
   };
 };
 
 export const deleteNode = (id) => {
   return {
     type: 'DELETE_NODE',
-    id
+    id,
   };
 };
 
 export const deleteAllNodes = () => {
   return {
-    type: 'DELETE_ALL_NODES'
+    type: 'DELETE_ALL_NODES',
   };
 };
 
 export const deleteSelectedNodes = () => {
   return {
-    type: 'DELETE_SELECTED_NODES'
+    type: 'DELETE_SELECTED_NODES',
   };
 };
 
 export const unlinkSelectedNodes = () => {
   return {
-    type: 'UNLINK_SELECTED_NODES'
+    type: 'UNLINK_SELECTED_NODES',
   };
 };
 
 export const selectNode = (id) => {
   return {
     type: 'SELECT_NODE',
-    id
+    id,
   };
 };
 
 export const deselectNodes = () => {
   return {
-    type: 'DESELECT_NODES'
+    type: 'DESELECT_NODES',
   };
 };
 
 export const selectAllNodes = () => {
   return {
-    type: 'SELECT_ALL_NODES'
+    type: 'SELECT_ALL_NODES',
   };
 };
 
-export const setNodeOsc1WaveType = (id, waveType) => {
-  return {
-    type: 'SET_NODE_OSC1_WAVE_TYPE',
-    id,
-    waveType
-  };
-};
-
-export const setNodeOsc2WaveType = (id, waveType) => {
-  return {
-    type: 'SET_NODE_OSC2_WAVE_TYPE',
-    id,
-    waveType
-  };
-};
 
 export const cloneNode = (id) => {
   return {
     type: 'CLONE_NODE',
-    id
-  };
-};
-
-export const setNodeAttack = (id, value) => {
-  return {
-    type: 'SET_NODE_ATTACK',
     id,
-    value
-  };
-};
-
-export const setNodeRelease = (id, value) => {
-  return {
-    type: 'SET_NODE_RELEASE',
-    id,
-    value
   };
 };
 
@@ -127,7 +72,7 @@ export const setNodeName = (id, name) => {
   return {
     type: 'SET_NODE_NAME',
     id,
-    name
+    name,
   };
 };
 
@@ -135,15 +80,7 @@ export const setNodeDisabledStatus = (id, status) => {
   return {
     type: 'SET_NODE_DISABLED_STATUS',
     id,
-    status
-  };
-};
-
-export const setNodePan = (id, pan) => {
-  return {
-    type: 'SET_NODE_PAN',
-    id,
-    pan
+    status,
   };
 };
 
@@ -151,7 +88,7 @@ export const linkNodes = (srcId, destId) => {
   return {
     type: 'LINK_NODES',
     srcId,
-    destId
+    destId,
   };
 };
 
@@ -159,14 +96,14 @@ export const unlinkNodes = (srcId, destId) => {
   return {
     type: 'UNLINK_NODES',
     srcId,
-    destId
+    destId,
   };
 };
 
 export const unlinkNode = (id) => {
   return {
     type: 'UNLINK_NODE',
-    id
+    id,
   };
 };
 
@@ -174,7 +111,7 @@ export const enqueueParticle = (id, particleId) => {
   return {
     type: 'ENQUEUE_PARTICLE',
     id,
-    particleId
+    particleId,
   };
 };
 
@@ -182,39 +119,39 @@ export const dequeueParticle = (id, particleId) => {
   return {
     type: 'DEQUEUE_PARTICLE',
     id,
-    particleId
+    particleId,
   };
 };
 
 export const dequeueParticles = () => {
   return {
-    type: 'DEQUEUE_PARTICLES'
+    type: 'DEQUEUE_PARTICLES',
   };
 };
 
 export const playNode = (id) => {
   return {
     type: 'PLAY_NODE',
-    id
+    id,
   };
 };
 
 export const stopNode = (id) => {
   return {
     type: 'STOP_NODE',
-    id
+    id,
   };
 };
 
 export const stopNodes = () => {
   return {
-    type: 'STOP_NODES'
+    type: 'STOP_NODES',
   };
 };
 
 export const stopSelectedNodes = () => {
   return {
-    type: 'STOP_SELECTED_NODES'
+    type: 'STOP_SELECTED_NODES',
   };
 };
 
@@ -222,7 +159,7 @@ export const setNodeLag = (id, lag) => {
   return {
     type: 'SET_NODE_LAG',
     id,
-    lag
+    lag,
   };
 };
 
@@ -230,39 +167,7 @@ export const setNodeProbability = (id, probability) => {
   return {
     type: 'SET_NODE_PROBABILITY',
     id,
-    probability
-  };
-};
-
-export const setNodeSendGain = (id, value) => {
-  return {
-    type: 'SET_NODE_SEND_GAIN',
-    id,
-    value
-  };
-};
-
-export const setNodeNoiseGain = (id, value) => {
-  return {
-    type: 'SET_NODE_NOISE_GAIN',
-    id,
-    value
-  };
-};
-
-export const setNodeOsc1Gain = (id, value) => {
-  return {
-    type: 'SET_NODE_OSC1_GAIN',
-    id,
-    value
-  };
-};
-
-export const setNodeOsc2Gain = (id, value) => {
-  return {
-    type: 'SET_NODE_OSC2_GAIN',
-    id,
-    value
+    probability,
   };
 };
 
@@ -270,7 +175,7 @@ export const setNodeVelocity = (id, value) => {
   return {
     type: 'SET_NODE_VELOCITY',
     id,
-    value
+    value,
   };
 };
 
@@ -278,7 +183,7 @@ export const setNodeNote = (id, value) => {
   return {
     type: 'SET_NODE_NOTE',
     id,
-    value
+    value,
   };
 };
 
@@ -286,7 +191,7 @@ export const setNodeChannel = (id, value) => {
   return {
     type: 'SET_NODE_CHANNEL',
     id,
-    value
+    value,
   };
 };
 
@@ -294,7 +199,7 @@ export const setNodeOctave = (id, value) => {
   return {
     type: 'SET_NODE_OCTAVE',
     id,
-    value
+    value,
   };
 };
 
@@ -302,7 +207,7 @@ export const updateSelectedNodePositionByDelta = (dx, dy) => {
   return {
     type: 'UPDATE_SELECTED_NODE_POSITION_BY_DELTA',
     dx,
-    dy
+    dy,
   };
 };
 
@@ -310,14 +215,14 @@ export const updateNodePositionByDelta = (dx, dy) => {
   return {
     type: 'UPDATE_NODE_POSITION_BY_DELTA',
     dx,
-    dy
+    dy,
   };
 };
 
 export const hydrateNodes = (payload) => {
   return {
     type: 'HYDRATE_NODES',
-    payload
+    payload,
   };
 };
 
@@ -325,6 +230,6 @@ export const setNodeMidiOutput = (id, outputId) => {
   return {
     type: 'SET_NODE_MIDI_OUTPUT',
     id,
-    outputId
+    outputId,
   };
 };
