@@ -1,6 +1,6 @@
 export const setSelection = (objType) => {
   return {
     type: 'SET_SELECTION',
-    objType
+    objType,
   };
 };
