@@ -2,7 +2,7 @@ export const addFreehandStream = (position, event) => {
   return {
     type: 'ADD_FREEHAND_STREAM',
     position,
-    event
+    event,
   };
 };
 
@@ -10,7 +10,7 @@ export const addCircularStream = (position, event) => {
   return {
     type: 'ADD_CIRCULAR_STREAM',
     position,
-    event
+    event,
   };
 };
 
@@ -18,33 +18,33 @@ export const addLinearStream = (position, event) => {
   return {
     type: 'ADD_LINEAR_STREAM',
     position,
-    event
+    event,
   };
 };
 
 export const deleteStream = (id) => {
   return {
     type: 'DELETE_STREAM',
-    id
+    id,
   };
 };
 
 export const deleteSelectedStreams = () => {
   return {
-    type: 'DELETE_STREAMS'
+    type: 'DELETE_STREAMS',
   };
 };
 
 export const deleteALLStreamS = () => {
   return {
-    type: 'DELETE_ALL_STREAMS'
+    type: 'DELETE_ALL_STREAMS',
   };
 };
 
 export const selectStream = (id) => {
   return {
     type: 'SELECT_STREAM',
-    id
+    id,
   };
 };
 
@@ -52,7 +52,7 @@ export const setStreamName = (id, name) => {
   return {
     type: 'SET_STREAM_NAME',
     id,
-    name
+    name,
   };
 };
 
@@ -60,7 +60,7 @@ export const setStreamDisabledStatus = (id, status) => {
   return {
     type: 'SET_STREAM_DISABLED_STATUS',
     id,
-    status
+    status,
   };
 };
 
@@ -68,7 +68,7 @@ export const setStreamSpeed = (id, speed) => {
   return {
     type: 'SET_STREAM_SPEED',
     id,
-    speed
+    speed,
   };
 };
 
@@ -76,7 +76,7 @@ export const setStreamCount = (id, count) => {
   return {
     type: 'SET_STREAM_COUNT',
     id,
-    count
+    count,
   };
 };
 
@@ -84,7 +84,7 @@ export const updateSelectedStreamPositionByDelta = (dx, dy) => {
   return {
     type: 'UPDATE_SELECTED_STREAM_POSITION_BY_DELTA',
     dx,
-    dy
+    dy,
   };
 };
 
@@ -92,19 +92,19 @@ export const updateStreamPositionByDelta = (dx, dy) => {
   return {
     type: 'UPDATE_STREAM_POSITION_BY_DELTA',
     dx,
-    dy
+    dy,
   };
 };
 
 export const hydrateStreams = (payload) => {
   return {
     type: 'HYDRATE_STREAMS',
-    payload
+    payload,
   };
 };
 
 export const deselectStreams = () => {
   return {
-    type: 'DESELECT_STREAMS'
+    type: 'DESELECT_STREAMS',
   };
 };
