@@ -1,12 +1,12 @@
 export const toggleTransport = () => {
   return {
-    type: 'TOGGLE_TRANSPORT'
+    type: 'TOGGLE_TRANSPORT',
   };
 };
 
 export const updateFPSCount = (count) => {
   return {
     type: 'UPDATE_FPS_COUNT',
-    count
+    count,
   };
 };
