@@ -2,19 +2,19 @@ export default {
   app: {
     collisionDistance: 22,
     doubleClickDistance: 20,
-    maxFrequency: 1500
+    maxFrequency: 1500,
   },
   transport: {
-    height: 35
+    height: 35,
   },
   canvas: {
-    backgroundColor: 'rgba(0, 0, 0, 0.0)'
+    backgroundColor: 'rgba(0, 0, 0, 0.0)',
   },
   particle: {
     count: 1,
   },
   midi: {
-    color: 'limegreen'
+    color: 'limegreen',
   },
   stream: {
     color: 'deeppink',
@@ -22,21 +22,21 @@ export default {
     size: 100,
     strokeStyle: 'gray',
     lineWidth: 2,
-    lineDash: [3, 30]
+    lineDash: [3, 30],
   },
   circularStream: {
     strokeStyle: 'gray',
     lineWidth: 2,
     lineDash: [3, 3],
     fillStyle: 'gold',
-    textAlign: 'center'
+    textAlign: 'center',
   },
   linearStream: {
     strokeStyle: 'gray',
     lineWidth: 2,
     lineDash: [3, 3],
     fillStyle: 'gold',
-    textAlign: 'center'
+    textAlign: 'center',
   },
   selectedStream: {
     strokeStyle: 'gold',
@@ -44,56 +44,19 @@ export default {
   link: {
     strokeStyle: 'gray',
     lineWidth: 3,
-    lineDash: [3, 3]
+    lineDash: [3, 3],
   },
   unlink: {
-    strokeStyle: 'crimson'
+    strokeStyle: 'crimson',
   },
   knob: {
-    zeroColor: 'dimgray'
+    zeroColor: 'dimgray',
   },
-  fader: {
-    marks: {
-      0: '-0',
-      0.1: '-1',
-      0.2: '-2',
-      0.3: '-3',
-      0.4: '-4',
-      0.5: '-5',
-      0.6: '-6',
-      0.7: '-7',
-      0.8: '-8',
-      0.9: '-9',
-      1: '-10'
-    }
-  },
-  waveToggle: {
-    emptyMarks: {
-      0: ' ',
-      1: ' ',
-      2: ' ',
-      3: ' '
-    }
-  },
+  
   controlPanel: {
     width: 300,
-    adsr: {
-      marks: {
-        0.0: '-',
-        0.1: ' ',
-        0.2: '-',
-        0.3: ' ',
-        0.4: '-',
-        0.5: ' ',
-        0.6: '-',
-        0.7: ' ',
-        0.8: '-',
-        0.9: ' ',
-        1.0: ' '
-      }
-    }
   },
   menu: {
-    width: 56
-  }
+    width: 56,
+  },
 };
